Flower_Image_Recognition_Modal/requirements.txt
torch
torchvision
timm
Pillow
matplotlib
seaborn
scikit-learn
numpy
pandas
tqdm
jupyter